requirements.txt
#requirements
fastapi
uvicorn[standard]
python-dotenv
httpx
pytest
pytest-asyncio
google-genai
fal-client
pymango
bcrypt
python-multipart
jinja2